bug 948151: Add SENTRY_ENVIRONMENT
SENTRY_ENVIRONMENT will be used to set the environment for Sentry
reporting, so that multiple environment can report to the same sentry
app instance. This would have been useful when we were transitioning
from MozMEAO to MozIT, and will be useful to distinguish between standby
and production, and between demo deploys and the official staging.