Refactoring VPC code
This is an exercise in decoupling the vpc creation from what was
generated by kops. Everything here was rewritten to move the VPC code
generated by kops to its own module.

This brings the VPC more inline to what a standard VPC layout should
look like and it gives us further flexibility moving forward when
creating AWS resources that requires a VPC.